Twitch: Contextual rate limiting. Discord, help and get random midi commands.

+++ app/lib/afhbot/modules/twitch-responses.rb
@@ -1,24 +1,86 @@
 module AFHBot
 
+  require 'open-uri'
+
   module TwitchResponses
 
     class Responses
 
-      @@allowed_methods = ['hello','ping']
+      @@allowed_methods = ['help', 'discord', 'midi']
 
       def self.allowed_method?(method_name)
-        return true if @@allowed_methods.include? method_name
-        false 
+        @@allowed_methods.include? method_name
+      end
+
+      def self.help(log, twitch, msg)
+        command_list = @@allowed_methods.map { |k|
+                            "#{twitch.config['command-prefix']}#{k}" }
+                            .join(' ')
+        twitch.msg(msg['to'], "List of commands: #{command_list}")
+      end
+
+      def self.discord(log, twitch, msg)
+        twitch.msg(msg['to'], "Link to Aviancer's Discord: https://discord.gg/WQHju3U")
+      end
+
+      def self.quote(log, twitch, msg)
+        # By, Quote, (Time)
+
+        # ie. allow for others to be quoted as well.
+        # Return #number for quote
+
+        # How do we keep bot from recording things off stream though?
+      end
+
+      def self.learn(log, twitch, msg)
+        # Learn definitions: name, definition
+
+        # Use another command to read them back.
+      end
+
+      def self.seen(log, twitch, msg)
+        # When was 'nick' last seen on channel
+
+        # Consider actual presence or just messaging?
+      end
+
+      # vgmusic.com random midi integration
+      def self.midi(log, twitch, msg)
+        twitch.add_rate_limiter(:RESPONSES_MIDI)
+        if twitch.rate_limited?(:RESPONSES_MIDI, 5, 10)
+          twitch.msg(msg['to'], "Midi Radio is on cooldown, wait a bit and try again.")
+          return
+        end
+
+        begin
+          doc = URI.open("https://www.vgmusic.com/cgi/random.cgi?") { |io| io.read() }
+        rescue => error
+          twitch.msg(msg['to'], "Midi Radio: Ran into an error, failed to get a random midi.")
+          log.error("AFHBot::TwitchResponses::#{__method__}: Failed to query vgmusic.com: #{error}")
+          return
+        end
+
+        patterns = {
+          :length => /Song Play Length:(:? (.{1,20}?))</m,
+          :url => /File URL is(:? (.{1,100}?))<br>/m,
+          :system => /Game System:(:? (.{1,30}?))<br>/m,
+          :game => /Game Name:(:? (.{1,50}?))<br>/m,
+          :title => /Song Title:(:? (.{1,50}?))<br>/m
+        }
+        
+        midi = patterns.map do |key, regexp|
+          result = doc.match(regexp)
+          # If match found, return match. Otherwise default.
+          [key, result ? result[1] : "Unknown #{key}"]
+        end.to_h
+
+        reply = "Midi Radio: #{midi[:url]} -> " \
+                 "'#{midi[:title]}' from '#{midi[:game]}' on " \
+                 "#{midi[:system]}. Length: #{midi[:length]}."
+        log.info("AFHBot::TwitchResponses::#{__method__} suggesting #{reply}")
+        twitch.msg(msg['to'], reply)
       end
  
-      def self.hello(twitch, msg)
-        twitch.msg(msg['to'], "Hello #{msg['from']}!")
-      end
-
-      def self.ping(twitch, msg)
-        twitch.msg(msg['to'], "Pong!")
-      end
-
     end
     
     def self.commands(log, moduleconfig)
@@ -39,17 +101,21 @@ module AFHBot
         }
         msg['to'], msg['text'] = twitch.parseparams_privmsg(event[:params])
 
-        # TODO: What if message isn't to a channel?
+        # Filter out messages not sent to channels in configuration
+        return if not twitch.allowed_channel?(msg['to'])
+
         if msg['text'].chars.first == moduleconfig["command-prefix"]
           msg['command'] = 
             msg['text'].delete_prefix(moduleconfig["command-prefix"]) # Strip command prefix -> method name
             .split(" ")
             .first
             .downcase
-          if AFHBot::TwitchResponses::Responses.allowed_method?(msg['command'])
-            AFHBot::TwitchResponses::Responses.send(msg['command'], twitch, msg)
+          if Responses.allowed_method?(msg['command'])
+            _logevent(log, __method__, msg, level: :info)
+            Responses.send(msg['command'], log, twitch, msg)
           else
-            log.debug("Unrecognized Twitch chat command: #{msg['command']}")
+            log.debug("Nick '#{msg['from']}' sent unrecognized Twitch chat command: #{msg['command']}")
+            _logevent(log, __method__, msg, level: :debug)
           end
         end
 
@@ -58,12 +124,12 @@ module AFHBot
 
     def self._logevent(log, method, event, level: :debug)
       # Max length willing to log of message is 512 characters.
-      message = event.options.inspect() # Will contain each event 'option' parameter
+      message = event.inspect() # Will contain each event 'option' parameter
       if message.length > 512
         message = message[0,512] + "..."
       end
       # send dynamically calls the method in log defined by 'level', eg. log.debug
-      log.send(level, "AFHBot::Group::#{method} event: srv(#{event.channel.server.name}) src(#{_event2author(event)}) msg(#{message})")
+      log.send(level, "AFHBot::TwitchResponses::#{method} event: chn(#{event['to']}) src(#{event['from']}) msg(#{message})")
     end
 
     def self._catcherror(log, method, event, msg, no_chat: false)
